feat: boton a reserva
