+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Peluquería"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivityHome"
+            android:name=".MainActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivityHistory"
@@ -25,7 +25,7 @@
             android:name=".MainActivityBook"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".MainActivityHome"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
