+++ app/src/main/res/layout/activity_main.xml
@@ -8,170 +8,4 @@
     android:background="@color/white"
     tools:context=".MainActivity">
 
-    <TextView
-        android:id="@+id/bienvenidaUsuario"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="center"
-        android:text="@string/texto_bienvenidaUsuario"
-        app:layout_constraintBottom_toTopOf="@+id/Titulo_reservas"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/Titulo_reservas"
-        android:layout_width="217dp"
-        android:layout_height="53dp"
-        android:layout_marginBottom="8dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="center"
-        android:text="@string/reservas_principal"
-        app:layout_constraintBottom_toTopOf="@+id/textViewMisdarosDeContacto"
-        app:layout_constraintEnd_toEndOf="@+id/bienvenidaUsuario"
-        app:layout_constraintStart_toStartOf="@+id/bienvenidaUsuario" />
-
-    <TextView
-        android:id="@+id/textViewMisdarosDeContacto"
-        android:layout_width="296dp"
-        android:layout_height="36dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="30dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="left"
-        android:text="@string/mis_daros_de_contacto"
-        app:layout_constraintBottom_toTopOf="@+id/textView3"
-        app:layout_constraintEnd_toStartOf="@+id/imageButtonEdit"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <ImageButton
-        android:id="@+id/imageButtonEdit"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:contentDescription="Este es el botón de edición"
-        android:hint="Este es el botón de edición"
-        android:background="@color/white"
-        app:layout_constraintBottom_toTopOf="@+id/textView3"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:srcCompat="@drawable/ic_edit" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="0dp"
-        android:layout_height="34dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="16dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="left"
-        android:text="@string/nombre_usuario"
-        app:layout_constraintBottom_toTopOf="@+id/textView4"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/imageView" />
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
-        android:layout_marginStart="15dp"
-        app:layout_constraintBottom_toTopOf="@+id/imageView2"
-        app:layout_constraintEnd_toStartOf="@+id/textView3"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_person_foreground" />
-
-    <TextView
-        android:id="@+id/textView4"
-        android:layout_width="0dp"
-        android:layout_height="34dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="15dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="left"
-        android:text="@string/telefonoContacto_usuario"
-        app:layout_constraintBottom_toTopOf="@+id/textViewMiReserva"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.058"
-        app:layout_constraintStart_toEndOf="@+id/imageView2" />
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
-        android:layout_marginStart="15dp"
-        android:layout_marginBottom="8dp"
-        app:layout_constraintBottom_toTopOf="@+id/textViewMiReserva"
-        app:layout_constraintEnd_toStartOf="@+id/textView4"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_call_foreground" />
-
-    <TextView
-        android:id="@+id/textViewMiReserva"
-        android:layout_width="227dp"
-        android:layout_height="40dp"
-        android:layout_marginEnd="50dp"
-        android:layout_marginBottom="8dp"
-        android:autoSizeTextType="uniform"
-        android:gravity="left"
-        android:text="Mi Reserva:"
-        app:layout_constraintBottom_toTopOf="@+id/scrollViewPrincipal"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.179"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <ScrollView
-        android:id="@+id/scrollViewPrincipal"
-        android:layout_width="0dp"
-        android:layout_height="250dp"
-        android:layout_marginBottom="8dp"
-        android:autoSizeTextType="uniform"
-        android:background="@color/white"
-        android:contentDescription="En este scrollView vemos las reservas"
-        android:gravity="center"
-        android:hint="En este scrollView vemos las reservas"
-        app:layout_constraintBottom_toTopOf="@+id/boton_NuevaReserva"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <androidx.cardview.widget.CardView
-                android:layout_width="match_parent"
-                android:layout_height="match_parent" />
-        </LinearLayout>
-    </ScrollView>
-
-    <Button
-        android:id="@+id/boton_NuevaReserva"
-        android:layout_width="171dp"
-        android:layout_height="68dp"
-        android:layout_marginBottom="8dp"
-        android:onClick="aReserva"
-        android:text="@string/boton_nuevaReserva"
-        android:textSize="18dp"
-        app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"
-        app:layout_constraintEnd_toEndOf="@+id/scrollViewPrincipal"
-        app:layout_constraintStart_toStartOf="@+id/scrollViewPrincipal" />
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottomNavigationView"
-        style="@style/Widget.MaterialComponents.BottomNavigationView.Colored"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:itemIconSize="50dp"
-        app:labelVisibilityMode="unlabeled"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:menu="@menu/bottom_menu" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
